requirements.txt
mysql-connector-python==8.0.26
pylint==2.11.1